Split up test_gp.py
