+++ .github/workflows/tests.yml
@@ -74,7 +74,10 @@ jobs:
 
           - |
             pymc/tests/distributions/test_timeseries.py
-            pymc/tests/test_gp.py
+            pymc/tests/gp/test_cov.py
+            pymc/tests/gp/test_gp.py
+            pymc/tests/gp/test_mean.py
+            pymc/tests/gp/test_util.py
             pymc/tests/test_model.py
             pymc/tests/test_model_graph.py
             pymc/tests/test_ode.py
@@ -148,7 +151,7 @@ jobs:
         test-subset:
           - pymc/tests/test_variational_inference.py pymc/tests/test_initial_point.py
           - pymc/tests/test_pickling.py pymc/tests/test_profile.py pymc/tests/test_step.py
-          - pymc/tests/test_gp.py pymc/tests/test_ode.py pymc/tests/test_smc.py pymc/tests/test_parallel_sampling.py
+          - pymc/tests/gp/test_cov.py pymc/tests/gp/test_gp.py pymc/tests/gp/test_mean.py pymc/tests/gp/test_util.py pymc/tests/test_ode.py pymc/tests/test_smc.py pymc/tests/test_parallel_sampling.py
           - pymc/tests/test_sampling.py pymc/tests/test_posteriors.py
 
       fail-fast: false
